Moved adapter from package module
