+++ frictionless/adapter/__init__.py
@@ -1,0 +1,2 @@
+from .adapter import Adapter
+from .mapper import Mapper
