+++ frictionless/adapter/adapter.py
@@ -1,17 +1,15 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, TypeVar, Generic
+from typing import TYPE_CHECKING, TypeVar, Generic, Optional, Any
 from ..dialect import Control
 
 if TYPE_CHECKING:
-    from .package import Package
     from ..catalog import Catalog
+    from ..package import Package
 
 ControlType = TypeVar("ControlType", bound=Control)
 
 
-# TODO: remove here and in the system after rebase to Adapter
 class Adapter(Generic[ControlType]):
-    # TODO: shall we accept aslo source?
     def __init__(self, control: ControlType):
         self.control = control
 
@@ -25,16 +23,16 @@ class Adapter(Generic[ControlType]):
 
     # Read
 
-    def read_catalog(self) -> Catalog:
+    def read_catalog(self) -> Optional[Catalog]:
         pass
 
-    def read_package(self) -> Package:
+    def read_package(self) -> Optional[Package]:
         pass
 
     # Write
 
-    def write_catalog(self, catalog: Catalog):
+    def write_catalog(self, catalog: Catalog) -> Optional[Any]:
         pass
 
-    def write_package(self, package: Package):
+    def write_package(self, package: Package) -> Optional[Any]:
         pass
