+++ frictionless/adapter/mapper.py
@@ -1,0 +1,2 @@
+class Mapper:
+    pass
