+++ frictionless/formats/sql/adapter.py
@@ -3,7 +3,8 @@ import re
 from typing import TYPE_CHECKING
 from urllib.parse import urlsplit, urlunsplit
 from .control import SqlControl
-from ...package import Package, Adapter
+from ...adapter import Adapter
+from ...package import Package
 from ...platform import platform
 from ...resource import Resource
 from .mapper import SqlMapper
