+++ frictionless/package/__init__.py
@@ -1,2 +1,1 @@
-from .adapter import Adapter
 from .package import Package
