+++ frictionless/plugin.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, Optional, List, Any, Type
 
 if TYPE_CHECKING:
     from .resource import Resource, Loader, Parser
-    from .package import Adapter
+    from .adapter import Adapter
     from .checklist import Check
     from .dialect import Control
     from .error import Error
